requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
cloudinary==1.15.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cssselect==1.0.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.4
django-autoslug==1.9.3
django-braces==1.13.0
django-cors-headers==2.5.2
django-cors-middleware==1.3.1
django-extensions==2.1.4
django-filter==2.1.0
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.1.0
django-simple-history==2.7.0
django-taggit==0.23.0
django-taggit-serializer==0.1.7
django-templated-mail==1.1.1
<<<<<<< HEAD
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
=======
djangorestframework==3.9.2
>>>>>>> 164857220-feature(Comments): Crud operations for comments
djoser==1.5.1
entrypoints==0.3
flake8==3.7.7
gunicorn==19.9.0
idna==2.8
isort==4.3.16
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.3.3
markdown2==2.3.7
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
more-itertools==7.0.0
oauthlib==3.0.1
openapi-codec==1.3.2
pathlib2==2.3.3
pbr==5.1.3
Pillow==5.4.1
# pkg-resources==0.0.0
pluggy==0.9.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
pyquery==1.4.0
pytest==4.0.1
pytest-cov==2.6.0
pytest-django==3.4.4
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.9
readtime==1.0.6
requests==2.21.0
requests-oauthlib==1.2.0
sendgrid==5.6.0
simplejson==3.16.0
six==1.10.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
soupsieve==1.9
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
